Add query optimization capabilities for some predicate pushdown in Neo4j Connector
Predicates that now can be pushed down include: AND, OR, IN, NOT, ADD(BIGINT), SUBTRACT(BIGINT),
EQUAL(BIGINT, INTEGER, DOUBLE, CHAR and VARCHAR), NOT EQUAL(BIGINT).
There are still some problems in the variable translation algorithm, and the query for the number of nodes
and relationships above 10 will not be executed normally.